refactor(runtime): extract fanout job and keyframe helpers from worker.c
Keep src/runtime/worker.c focused on worker lifecycle and the worker
thread loop by moving non-lifecycle helpers out:

- add src/runtime/fanout_job.{c,h} for sfu_worker_handle_fanout_job()
- add src/pipeline/keyframe.{c,h} for sfu_worker_request_keyframe_throttled()
- update pipeline/router.c, pipeline/ingress.c, and worker protocol tests to
  use the new headers
- register the new translation units in CMakeLists.txt

No behavior change; this only narrows worker.c's responsibilities and makes
the module boundaries match the pipeline split.
