+++ CMakeLists.txt
@@ -84,7 +84,9 @@ add_library(sfu_core STATIC
   src/pipeline/ingress.c
   src/pipeline/router.c
   src/pipeline/egress.c
+  src/pipeline/keyframe.c
   src/runtime/cpu.c
+  src/runtime/fanout_job.c
   src/runtime/timer.c
   src/runtime/fanout.c
   src/runtime/worker.c
